==> 01-scRNAseq Acute Wounds/05_subClustering/s5-Lymphoid/s1_clean_LymphoidCells.Rmd <==
---
title: "Sub-clustering of immune cells"
author: "Zhuang Liu"
date: "30/11/2021 Updated: `r Sys.Date()`"
output:
  html_document:
    toc: true #add a table of contents (TOC)
    toc_depth: 3 #headers' level in the toc
    toc_float: true #toc visible even when scrolled
    theme: lumen
    highlight: tango
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.align='center', fig.width = 12, warning = FALSE, message = FALSE)
```

```{r warning=FALSE, message=FALSE}
library(Seurat)
library(SeuratObject)
library(SeuratDisk)
library(tidyverse)
library(magrittr)
library(ggplot2)
library(patchwork)
library(harmony)
```

####################################################
# Step 1. Extract cells
####################################################
## 1.1 Load the whole cell types
```{r fig.width=8, fig.height=6}
hswound.combined.sct <- readRDS("00-tmpRawfiles/s1_ori_immu_lymphoid.rds") 

DimPlot(object = hswound.combined.sct, 
        reduction = "umap", 
        label = TRUE, 
        group.by = "SCT_snn_res.1", label.size = 3) + NoAxes() + NoLegend()
DimPlot(object = hswound.combined.sct, 
        reduction = "umap", 
        label = TRUE, 
        group.by = "MainCellTypes", label.size = 3) + NoAxes() + NoLegend()
```

## 1.2 Rerun the SCT and clustering (all Conditions)
```{r}
table(hswound.combined.sct$SCT_snn_res.1)
hswound.combined.sct$fill <- ifelse(hswound.combined.sct$SCT_snn_res.1 %in% c(16, 17), "Yes", "No")
hswound.immu <- subset(x = hswound.combined.sct, subset = fill == "No")
hswound.immu@meta.data <- droplevels(hswound.immu@meta.data)

DimPlot(object = hswound.immu, 
        reduction = "umap", 
        label = TRUE, 
        group.by = "SCT_snn_res.1", label.size = 3) + NoAxes() + NoLegend()
DimPlot(object = hswound.immu, 
        reduction = "umap", 
        label = TRUE, 
        group.by = "MainCellTypes", label.size = 3) + NoAxes() + NoLegend()

rm(hswound.combined.sct);invisible(gc())

DefaultAssay(hswound.immu) <- "RNA"
hswound.immu <- DietSeurat(hswound.immu, counts = TRUE, data = TRUE, scale.data = FALSE,
                    assays = "RNA", dimreducs = NULL)
#Remove unneeded columns
hswound.immu@active.ident <- hswound.immu$orig.ident
hswound.immu$SCT_snn_res.0.5 <- NULL
hswound.immu$SCT_snn_res.0.8 <- NULL
hswound.immu$SCT_snn_res.1.2 <- NULL
hswound.immu$SCT_snn_res.1 <- NULL
hswound.immu$seurat_clusters <- NULL
hswound.immu$nCount_SCT <- NULL
hswound.immu$nFeature_SCT <- NULL
hswound.immu$fill <- NULL

hswound.list <- SplitObject(hswound.immu, split.by = "orig.ident")
rm(hswound.immu);invisible(gc())
hswound.list <- lapply(X = hswound.list, FUN = function(x) {
  x <- NormalizeData(x)
  x <- CellCycleScoring(object = x, g2m.features = cc.genes$g2m.genes, s.features = cc.genes$s.genes, set.ident = FALSE)
  x$CC.Difference <- x$S.Score - x$G2M.Score
  x <- SCTransform(x, vars.to.regress = c("percent.mt", "CC.Difference"), variable.features.n = 3000, verbose = FALSE, return.only.var.genes = TRUE, method = "glmGamPoi")
})

vg = lapply(hswound.list, VariableFeatures)
t = table(unlist(vg))

vg$all = SelectIntegrationFeatures(hswound.list)
source("../Functions/overlap_phyper2.R")

vg.all = names(t)[t > 1]
length(vg.all)

# filter to only have genes detected in all sections.
rn = lapply(hswound.list, rownames)
nE = colSums(Reduce(rbind, lapply(rn, function(x) vg.all %in% x)))
vg.all = vg.all[nE == length(hswound.list)]
length(vg.all)

vg$finalall <- vg.all

features.all <- SelectIntegrationFeatures(object.list = hswound.list, nfeatures = 2000)
rn <- lapply(hswound.list, rownames) 
nE <- colSums(Reduce(rbind, lapply(rn, function(x) features.all %in% x))) 
features.all <- features.all[nE == length(hswound.list)]

vg$featuresall <- features.all
overlap_phyper2(vg, vg)

####----Prepare the integration----####
hswound.list <- PrepSCTIntegration(object.list = hswound.list, anchor.features = features.all)

hswound.immu.com <- merge(hswound.list[[1]], 
                         y = hswound.list[2:length(hswound.list)],
                         project = "woundimmu",
                         merge.data = TRUE)

VariableFeatures(hswound.immu.com) <- features.all
hswound.immu.com <- RunPCA(object = hswound.immu.com, assay = "SCT", npcs = 50, features = VariableFeatures(hswound.immu.com))
ElbowPlot(hswound.immu.com, ndims = 40, reduction = "pca")

hswound.immu.com <- RunHarmony(object = hswound.immu.com,
                                    assay.use = "SCT",
                                    reduction = "pca",
                                    dims.use = 1:20,
                                    group.by.vars = "orig.ident",
                                    plot_convergence = TRUE)

#rn pca, umap, clustering
hswound.immu.com <- FindNeighbors(object = hswound.immu.com, assay = "SCT", reduction = "harmony", dims = 1:20, k.param = 40)
hswound.immu.com <- FindClusters(object = hswound.immu.com, resolution = 0.8)
hswound.immu.com <- FindClusters(object = hswound.immu.com, resolution = 0.5)
hswound.immu.com <- FindClusters(object = hswound.immu.com, resolution = 1.0)
hswound.immu.com <- FindClusters(object = hswound.immu.com, resolution = 1.2)

#hswound.immu.com <- RunUMAP(object = hswound.immu.com, assay = "SCT", reduction = "harmony", dims = #1:20, n.neighbors = 40, min.dist = 0.3, n.epochs = 500, spread = 1, learning.rate = 1)
hswound.immu.com <- RunUMAP(object = hswound.immu.com, assay = "SCT", reduction = "harmony", dims = 1:20, n.neighbors = 40, min.dist = 0.5)

hswound.immu.com$orig.ident <- factor(x = hswound.immu.com$orig.ident, levels =
                                            c("PWH26D0", "PWH26D1", "PWH26D7","PWH26D30",
                                              "PWH27D0", "PWH27D1", "PWH27D7","PWH27D30",
                                              "PWH28D0", "PWH28D1", "PWH28D7","PWH28D30"))
hswound.immu.com$Condition <- factor(x = hswound.immu.com$Condition, levels =
                                            c("Skin", "Wound1", "Wound7","Wound30"))

DimPlot(object = hswound.immu.com, 
        reduction = "umap", 
        label = F, 
        group.by = "orig.ident", label.size = 4.5) + NoAxes() 

table(hswound.immu.com$SCT_snn_res.0.5)
DimPlot(object = hswound.immu.com, 
        reduction = "umap", 
        label = TRUE, 
        group.by = "SCT_snn_res.0.5", label.size = 4.5) + NoAxes() 

table(hswound.immu.com$SCT_snn_res.0.8)
DimPlot(object = hswound.immu.com, 
        reduction = "umap", 
        label = TRUE, 
        group.by = "SCT_snn_res.0.8", label.size = 4.5) + NoAxes()

table(hswound.immu.com$SCT_snn_res.1)
DimPlot(object = hswound.immu.com, 
        reduction = "umap", 
        label = TRUE, 
        group.by = "SCT_snn_res.1", label.size = 4.5) + NoAxes()

table(hswound.immu.com$SCT_snn_res.1.2)
DimPlot(object = hswound.immu.com, 
        reduction = "umap", 
        label = TRUE, 
        group.by = "SCT_snn_res.1.2", label.size = 4.5) + NoAxes()

DimPlot(object = hswound.immu.com, 
        reduction = "umap", 
        label = TRUE, 
        group.by = "MainCellTypes", label.size = 5) + NoAxes() + ggtitle("")

DimPlot(hswound.immu.com, reduction = "umap", pt.size = .001,group.by = "SCT_snn_res.1", split.by = 'Condition', ncol = 4, label = F) + NoLegend()+ ggtitle("")
DimPlot(hswound.immu.com, reduction = "umap", pt.size = .001,group.by = "MainCellTypes", split.by = 'Condition', ncol = 4, label = F) + NoLegend()+ ggtitle("")

FeaturePlot(hswound.immu.com, features = c("MKI67", "CCR7"))

# Determine metrics to plot present in hswound.krt.com@meta.data
metrics <-  c("nCount_RNA", "nFeature_RNA", "S.Score", "G2M.Score", "percent.mt")

p3 <- FeaturePlot(
  hswound.immu.com,
  features = metrics,
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p3, function(x) {x + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 3)

VlnPlot(hswound.immu.com, features = metrics, group.by='SCT_snn_res.1', pt.size = 0, ncol = 2)

ggplot(hswound.immu.com@meta.data, aes(x = SCT_snn_res.0.8, fill = Condition, alpha = Patient)) +
    geom_bar(position = "fill")

library(future)
plan("multisession", workers = 4)
options(future.globals.maxSize = 15000 * 1024^2)

hswound.immu.com <- readRDS("s1_clean_immu_lymphoid.rds")
####----Change to RNA assay to prepare the DE analysis----####
DefaultAssay(hswound.immu.com) <- "RNA"
Idents(hswound.immu.com) <- "SCT_snn_res.0.8" #marker genes with cluster number
####----FindAllMarkers for all clusters----####
markers <- FindAllMarkers(
  hswound.immu.com,
  only.pos = TRUE,
  min.pct = 0.25,  #min.pct = 0.5, logfc.threshold = 0.5,
  logfc.threshold = 0.25,
  test.use = "MAST") #MAST has good FDR control and is faster than DESeq2, default test is wilcox

####----add the gene description to the cluster marker genes----####
annot <- readRDS("../Functions/humanAnnotation.rds")
####----add the info and reorder the columns----####
markers_f <- markers %>% left_join(., annot, by=c("gene" = "external_gene_name")) %>% 
  dplyr::select(6, 7, 2:4, 1, 5, everything()) %>% arrange(cluster, desc(avg_log2FC))
#marker genes with cluster number
data.table::fwrite(markers_f, file="s1_clean_immu_lymphoid_mg_res0.8.txt", sep="\t") 

hswound.immu.com <- NormalizeData(hswound.immu.com, verbose = TRUE,  normalization.method = "LogNormalize", scale.factor = 10000)

saveRDS(hswound.immu.com, file = "s1_clean_immu_lymphoid.rds")
```

```{r fig.height=11}
FeaturePlot(
  hswound.immu.com,
  features=c('CD14', 'S100A9', 'FCGR3A', 'IL1B', 'CD163'), 
  combine=T, cols = c("grey","red")
)

plotcom <- FeaturePlot(
  hswound.immu.com,
  features=c('CD8A', 'CD8B', 'CD4', 'CD40LG', 'FOXP3', 'TIGIT', 'CTLA4'), 
  combine=F, cols = c("grey","red")
)

# apply theme to each feature plot
for(i in 1:length(plotcom)){
  plotcom[[i]] <- plotcom[[i]] + NoLegend() + NoAxes()
}

cowplot::plot_grid(plotlist = plotcom, ncol = 4)
```

# SessionInfo
```{r}
sessionInfo()
```
